Move build utils to upper package
